deploy/databricks/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pydantic==2.9.2
numpy==1.26.4
databricks-sdk==0.38.0
